requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
google-cloud-aiplatform==1.72.0
python-dotenv==1.0.1
rocketride
